Improve sdk bundle size for SSR purposes
