+++ packages/sdk/package.json
@@ -74,8 +74,6 @@
     "@audius/fixed-decimal": "0.1.1",
     "@audius/hedgehog": "3.0.0-alpha.1",
     "@audius/spl": "2.0.0",
-    "@babel/core": "7.23.7",
-    "@babel/plugin-proposal-class-static-block": "7.21.0",
     "@babel/runtime": "7.18.3",
     "@improbable-eng/grpc-web-node-http-transport": "0.15.0",
     "@noble/hashes": "1.3.0",
@@ -88,33 +86,19 @@
     "async-mutex": "0.5.0",
     "async-retry": "1.3.1",
     "axios": "0.19.2",
-    "bn.js": "5.2.1",
     "borsh": "0.4.0",
-    "browser-or-node": "^2.1.1",
-    "bs58": "4.0.1",
-    "crc-32": "1.2.2",
     "cross-fetch": "4.0.0",
-    "esm": "3.2.25",
     "file-type": "16.5.3",
     "form-data": "3.0.0",
     "hashids": "2.2.10",
-    "interface-blockstore": "2.0.3",
-    "interface-store": "2.0.2",
-    "ipfs-unixfs-importer": "9.0.6",
     "isomorphic-ws": "5.0.0",
-    "keccak256": "1.0.2",
     "lodash": "4.17.21",
     "micro-aes-gcm": "0.4.0",
     "multiformats": "13.3.1",
     "node-abort-controller": "3.1.1",
     "node-localstorage": "1.3.1",
-    "rollup-plugin-shim": "1.0.0",
-    "safe-buffer": "5.2.1",
-    "secp256k1": "4.0.2",
     "semver": "6.3.0",
     "snakecase-keys": "5.4.5",
-    "stream-browserify": "3.0.0",
-    "token-types": "4.2.1",
     "type-fest": "4.26.1",
     "ulid": "2.3.0",
     "url": "0.11.1",
@@ -124,6 +108,8 @@
     "zod": "3.21.4"
   },
   "devDependencies": {
+    "@babel/core": "7.23.7",
+    "@babel/plugin-proposal-class-static-block": "7.21.0",
     "@babel/plugin-transform-runtime": "7.18.2",
     "@babel/preset-env": "7.16.11",
     "@babel/preset-typescript": "7.23.3",
@@ -139,7 +125,6 @@
     "@tsconfig/strictest": "2.0.5",
     "@types/async-retry": "1.4.4",
     "@types/bn.js": "5.1.0",
-    "@types/bs58": "4.0.1",
     "@types/eth-sig-util": "2.1.1",
     "@types/expect": "24.3.0",
     "@types/form-data": "2.5.0",
@@ -155,6 +140,7 @@
     "commander": "9.2.0",
     "crypto-browserify": "^3.12.0",
     "eslint-plugin-mocha": "10.0.3",
+    "esm": "3.2.25",
     "ethereumjs-wallet": "^1.0.2",
     "mocha": "9.2.2",
     "msw": "1.3.1",
@@ -167,8 +153,10 @@
     "rollup-plugin-dts": "4.2.0",
     "rollup-plugin-ignore": "1.0.10",
     "rollup-plugin-polyfill-node": "0.9.0",
+    "rollup-plugin-shim": "1.0.0",
     "rollup-plugin-terser": "7.0.2",
     "sinon": "9.0.2",
+    "stream-browserify": "3.0.0",
     "ts-node": "10.9.2",
     "tslib": "^2.6.2",
     "tsx": "^4.18.0",
